requirements.txt
pyautogui
types-PyAutoGUI
python-dotenv
